Added in support for new broadcast cmd in SPJS.  This closes #1
Instead of sending a GCODE comment (which no one really sees anyway)
this will now send out a broadcast command to the SPJS.  John added a
new widget which will watch for this broadcast and show the currently
driven axis and speed, as well as announce it with a wav file.

The broadcast command was added in the 1.80 version of SPJS.  I tested
it with older releases and they gracefully ignore the bad command.